+++ CHANGELOG.md
@@ -4,6 +4,10 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+### Changed
+- Subscriptions are no longer run asynchronously during catchup
+
 ## [2.0.22] - 2019-09-24
 ### Changed
 - Catchup verification logging now runs in MDC context